test: 소수 덧셈 기능 구현
- 정확한 소수 덧셈 결과를 반환하기 위해 BigDecimal 타입 도입
- 기존 int 타입의 숫자를 BigDecimal 타입의 숫자로 변경

+++ src/main/java/calculator/view/OutputView.java
@@ -1,5 +1,7 @@
 package calculator.view;
 
+import java.math.BigDecimal;
+
 public class OutputView {
 
     private static OutputView instance;
@@ -14,7 +16,7 @@ public class OutputView {
         return instance;
     }
 
-    public void renderingOutputView(int result) {
+    public void renderingOutputView(BigDecimal result) {
         System.out.println("결과 : " + result);
     }
 }
